- Changed the output of the OPL emulator from 32-bit integers to 32-bit   floats, so I can write its output directly to the stream buffer. In   addition, this lets me bring the OPL volume level much closer to the   standard MIDI volume.
SVN r863 (trunk)

+++ src/oplsynth/fmopl.h
@@ -38,7 +38,7 @@ void YM3812ResetChip(int which);
 int  YM3812Write(int which, int a, int v);
 unsigned char YM3812Read(int which, int a);
 int  YM3812TimerOver(int which, int c);
-void YM3812UpdateOne(int which, INT32 *buffer, int length);
+void YM3812UpdateOne(int which, float *buffer, int length);
 
 void YM3812SetTimerHandler(int which, OPL_TIMERHANDLER TimerHandler, int channelOffset);
 void YM3812SetIRQHandler(int which, OPL_IRQHANDLER IRQHandler, int param);
